Introduce analyze command to extract field names
When working with converting large numbers of Sigma rules, being
able to check whether a query's results contains the fields that are
searched for in the converted rule is integral to validating the query
will work correctly. This introduces a new analyze command, `fields`
which is provided a backend, any number of pipelines, and Sigma files,
and returns a list of unique fields that appear after the pipeline
transformations are taken into account.